+++ src/bespokelabs/curator/prompter/prompt_formatter.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, Dict, Optional, Type, TypeVar, Union
 
 from pydantic import BaseModel
 
-from bespokelabs.curator.request_processor.generic_request import GenericRequest
+from bespokelabs.curator.types.generic_request import GenericRequest
 
 T = TypeVar("T")
