+++ src/bespokelabs/curator/request_processor/base_online_request_processor.py
@@ -14,7 +14,8 @@ import resource
 from bespokelabs.curator.dataset import Dataset
 from bespokelabs.curator.request_processor.base_request_processor import BaseRequestProcessor
 from bespokelabs.curator.prompter.prompter import PromptFormatter
-from bespokelabs.curator.request_processor.generic_request import GenericRequest
+from bespokelabs.curator.status_tracker.online_status_tracker import OnlineStatusTracker
+from bespokelabs.curator.types.generic_request import GenericRequest
 from bespokelabs.curator.request_processor.event_loop import run_in_event_loop
 from bespokelabs.curator.request_processor.generic_response import GenericResponse
 import aiofiles
@@ -24,79 +25,6 @@ logger.setLevel(logging.INFO)
 
 DEFAULT_REQUESTS_PER_MINUTE = 10
 DEFAULT_TOKENS_PER_MINUTE = 1_000
-
-
-@dataclass
-class StatusTracker:
-    """Tracks the status of all requests."""
-
-    num_tasks_started: int = 0
-    num_tasks_in_progress: int = 0
-    num_tasks_succeeded: int = 0
-    num_tasks_failed: int = 0
-    num_tasks_already_completed: int = 0
-    num_api_errors: int = 0
-    num_other_errors: int = 0
-    num_rate_limit_errors: int = 0
-    available_request_capacity: float = 0
-    available_token_capacity: float = 0
-    last_update_time: float = field(default_factory=time.time)
-    max_requests_per_minute: int = 0
-    max_tokens_per_minute: int = 0
-    pbar: tqdm = field(default=None)
-    response_cost: float = 0
-    time_of_last_rate_limit_error: float = field(default=None)
-
-    def __str__(self):
-        return (
-            f"Tasks - Started: {self.num_tasks_started}, "
-            f"In Progress: {self.num_tasks_in_progress}, "
-            f"Succeeded: {self.num_tasks_succeeded}, "
-            f"Failed: {self.num_tasks_failed}, "
-            f"Already Completed: {self.num_tasks_already_completed}\n"
-            f"Errors - API: {self.num_api_errors}, "
-            f"Rate Limit: {self.num_rate_limit_errors}, "
-            f"Other: {self.num_other_errors}, "
-            f"Total: {self.num_other_errors + self.num_api_errors + self.num_rate_limit_errors}"
-        )
-
-    def update_capacity(self):
-        """Update available capacity based on time elapsed"""
-        current_time = time.time()
-        seconds_since_update = current_time - self.last_update_time
-
-        self.available_request_capacity = min(
-            self.available_request_capacity
-            + self.max_requests_per_minute * seconds_since_update / 60.0,
-            self.max_requests_per_minute,
-        )
-
-        self.available_token_capacity = min(
-            self.available_token_capacity
-            + self.max_tokens_per_minute * seconds_since_update / 60.0,
-            self.max_tokens_per_minute,
-        )
-
-        self.last_update_time = current_time
-
-    def has_capacity(self, token_estimate: int) -> bool:
-        """Check if there's enough capacity for a request"""
-        self.update_capacity()
-        has_capacity = (
-            self.available_request_capacity >= 1 and self.available_token_capacity >= token_estimate
-        )
-        if not has_capacity:
-            logger.debug(
-                f"No capacity for request with {token_estimate} tokens. "
-                f"Available capacity: {self.available_token_capacity} tokens, "
-                f"{self.available_request_capacity} requests."
-            )
-        return has_capacity
-
-    def consume_capacity(self, token_estimate: int):
-        """Consume capacity for a request"""
-        self.available_request_capacity -= 1
-        self.available_token_capacity -= token_estimate
 
 
 @dataclass
@@ -232,7 +160,7 @@ class BaseOnlineRequestProcessor(BaseRequestProcessor, ABC):
 
         # Initialize trackers
         queue_of_requests_to_retry: asyncio.Queue[APIRequest] = asyncio.Queue()
-        status_tracker = StatusTracker()
+        status_tracker = OnlineStatusTracker()
 
         # Get rate limits
         rate_limits = self.get_rate_limits()
@@ -427,7 +355,7 @@ class BaseOnlineRequestProcessor(BaseRequestProcessor, ABC):
         session: aiohttp.ClientSession,
         retry_queue: asyncio.Queue,
         save_filepath: str,
-        status_tracker: StatusTracker,
+        status_tracker: OnlineStatusTracker,
     ) -> None:
         """Common wrapper for handling a single request with error handling and retries.
 
@@ -439,7 +367,7 @@ class BaseOnlineRequestProcessor(BaseRequestProcessor, ABC):
             session (aiohttp.ClientSession): Async HTTP session
             retry_queue (asyncio.Queue): Queue for failed requests
             save_filepath (str): Path to save responses
-            status_tracker (StatusTracker): Tracks request status
+            status_tracker (OnlineStatusTracker): Tracks request status
         """
         try:
             generic_response = await self.call_single_request(
@@ -493,7 +421,7 @@ class BaseOnlineRequestProcessor(BaseRequestProcessor, ABC):
         self,
         request: APIRequest,
         session: aiohttp.ClientSession,
-        status_tracker: StatusTracker,
+        status_tracker: OnlineStatusTracker,
     ) -> GenericResponse:
         """Make a single API request without error handling.
 
@@ -503,7 +431,7 @@ class BaseOnlineRequestProcessor(BaseRequestProcessor, ABC):
         Args:
             request (APIRequest): Request to process
             session (aiohttp.ClientSession): Async HTTP session
-            status_tracker (StatusTracker): Tracks request status
+            status_tracker (OnlineStatusTracker): Tracks request status
 
         Returns:
             GenericResponse: The response from the API call
