+++ src/bespokelabs/curator/status_tracker/online_status_tracker.py
@@ -1,0 +1,80 @@
+import logging
+import time
+from dataclasses import dataclass, field
+
+import tqdm
+
+logger = logging.getLogger(__name__)
+
+
+@dataclass
+class OnlineStatusTracker:
+    """Tracks the status of all requests."""
+
+    num_tasks_started: int = 0
+    num_tasks_in_progress: int = 0
+    num_tasks_succeeded: int = 0
+    num_tasks_failed: int = 0
+    num_tasks_already_completed: int = 0
+    num_api_errors: int = 0
+    num_other_errors: int = 0
+    num_rate_limit_errors: int = 0
+    available_request_capacity: float = 0
+    available_token_capacity: float = 0
+    last_update_time: float = field(default_factory=time.time)
+    max_requests_per_minute: int = 0
+    max_tokens_per_minute: int = 0
+    pbar: tqdm = field(default=None)
+    response_cost: float = 0
+    time_of_last_rate_limit_error: float = field(default=None)
+
+    def __str__(self):
+        return (
+            f"Tasks - Started: {self.num_tasks_started}, "
+            f"In Progress: {self.num_tasks_in_progress}, "
+            f"Succeeded: {self.num_tasks_succeeded}, "
+            f"Failed: {self.num_tasks_failed}, "
+            f"Already Completed: {self.num_tasks_already_completed}\n"
+            f"Errors - API: {self.num_api_errors}, "
+            f"Rate Limit: {self.num_rate_limit_errors}, "
+            f"Other: {self.num_other_errors}, "
+            f"Total: {self.num_other_errors + self.num_api_errors + self.num_rate_limit_errors}"
+        )
+
+    def update_capacity(self):
+        """Update available capacity based on time elapsed"""
+        current_time = time.time()
+        seconds_since_update = current_time - self.last_update_time
+
+        self.available_request_capacity = min(
+            self.available_request_capacity
+            + self.max_requests_per_minute * seconds_since_update / 60.0,
+            self.max_requests_per_minute,
+        )
+
+        self.available_token_capacity = min(
+            self.available_token_capacity
+            + self.max_tokens_per_minute * seconds_since_update / 60.0,
+            self.max_tokens_per_minute,
+        )
+
+        self.last_update_time = current_time
+
+    def has_capacity(self, token_estimate: int) -> bool:
+        """Check if there's enough capacity for a request"""
+        self.update_capacity()
+        has_capacity = (
+            self.available_request_capacity >= 1 and self.available_token_capacity >= token_estimate
+        )
+        if not has_capacity:
+            logger.debug(
+                f"No capacity for request with {token_estimate} tokens. "
+                f"Available capacity: {self.available_token_capacity} tokens, "
+                f"{self.available_request_capacity} requests."
+            )
+        return has_capacity
+
+    def consume_capacity(self, token_estimate: int):
+        """Consume capacity for a request"""
+        self.available_request_capacity -= 1
+        self.available_token_capacity -= token_estimate
