+++ src/bespokelabs/curator/types/generic_request.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional, Type
+from typing import Any, Dict, List
 
 from pydantic import BaseModel
